pequeñas reorganizaciones del código. subimos archivos de sonido
